Homework-3: in case of page re-fresh pages routes to login or courses, depending in user logged in
