+++ src/components/Container/Container.jsx
@@ -2,14 +2,30 @@ import './Container.css';
 
 import React, { useEffect } from 'react';
 import { Outlet, useNavigate } from 'react-router-dom';
+import { useDispatch } from 'react-redux';
 import Header from '../Header/Header.jsx';
+import { loginUser } from '../../store/user/actionCreators';
 
 function Container() {
 	const navigate = useNavigate();
+	const dispatch = useDispatch();
 
 	useEffect(() => {
-		navigate('/login');
+		let auth = JSON.parse(localStorage.getItem('authData'));
+		if (auth !== null) {
+			dispatch(
+				loginUser({
+					name: auth.name,
+					email: auth.email,
+					token: auth.token,
+				})
+			);
+			navigate('/courses');
+		} else {
+			navigate('/login');
+		}
 	}, []);
+
 	return (
 		<div className='container'>
 			<Header />
